Have addModFinalizer expose the local type environment.
Summary:
This annotates the splice point with 'HsSpliced fins e' where 'e' is the
result of the splice. 'fins' are the finalizers that the typechecker will add to
the global environment while exposing the current local environment to
them.

If 'reify' is called in a finalizer, it is able to find all names in scope at the
point where 'addModFinalizer' was called.

Test Plan: ./validate

Reviewers: simonpj, goldfire, austin, bgamari

Subscribers: thomie, mboes

Differential Revision: https://phabricator.haskell.org/D2286

GHC Trac Issues: #11832